second commit

+++ app/Http/Controllers/BillController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Bill;
 use App\Models\Customer;
+use App\Models\BillItem;
 use Illuminate\Http\Request;
 
 class BillController extends Controller
@@ -16,6 +17,7 @@ class BillController extends Controller
     public function index()
     {
         //
+        $this->authorize('viewAny', Bill::class);
         $bills=Bill::all();
         return view('bill.index', compact('bills'));
     }
@@ -28,6 +30,7 @@ class BillController extends Controller
     public function create()
     {
         //
+        $this->authorize('create', Bill::class);
         $customers=Customer::all();
         return view('bill.create',compact('customers'));
     }
@@ -41,8 +44,37 @@ class BillController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create', Bill::class);
         $input=$request->all();
-        $bill=Bill::create($input);
+        $customer=Customer::find($input['customer_id']);
+        $tarrif=$customer->tarrif; 
+        $biiItems=array();
+        $billItems[]=['narrative'=>'Consumption in cubic meters', 'quantity'=>$input['consumption'], 'unit'=>$tarrif->price, 'amount'=>$input['consumption'] * $tarrif->price, 'status'=>'Active'];
+        foreach($tarrif->tarrifItems as $item){
+            if($item->type=='Fixed Amount'){
+                $quantity=1;
+                $unit=$item->amount;
+                $amount=$quantity * $unit;
+                
+            }elseif($item->type == 'Amount per cubic meter'){
+                $quantity=$input['consumption'];
+                $unit=$item->amount;
+                $amount=$quantity * $unit;
+
+            }elseif($item->type =='% of consumption'){
+                $quantity=$item->amount/100;
+                $unit=$input['consumption'] * $tarrif->price;
+                $amount=$quantity * $unit;
+            }
+        $billItems[]=['narrative'=>$item->name, 'quantity'=>$quantity, 'unit'=>$unit, 'amount'=>$amount, 'status'=>'Active'];
+        }
+         $total=array_sum(array_column($billItems,'amount'));
+         $billD=['customer_id'=>$input['customer_id'], 'start_date'=>$input['start_date'], 'end_date'=>$input['end_date'], 'date'=>date('Y-m-d'), 'status'=>'Unpaid', 'balance'=>$total, 'amount'=>$total];
+        $bill=Bill::create($billD);
+        foreach($billItems as $item){
+            $item['bill_id']=$bill->id;
+            $billItem=BillItem::create($item);
+        }
         return redirect('/bill/'.$bill->id);
     }
 
@@ -55,6 +87,7 @@ class BillController extends Controller
     public function show(Bill $bill)
     {
         //
+        $this->authorize('view', $bill);
         return view('bill.show', compact('bill'));
     }
 
@@ -67,6 +100,7 @@ class BillController extends Controller
     public function edit(Bill $bill)
     {
         //
+        $this->authorize('update', $bill);
         $customers=Customer::all();
         return view('bill.edit', compact('bill','customers'));
     }
@@ -81,6 +115,7 @@ class BillController extends Controller
     public function update(Request $request, Bill $bill)
     {
         //
+        $this->authorize('update', $bill);
         $input=$request->all();
         $bill->update($input);
         return redirect('/bill/'.$bill->id);
@@ -95,6 +130,7 @@ class BillController extends Controller
     public function destroy(Bill $bill)
     {
         //
+        $this->authorize('delete', $bill);
         $bill->delete();
         return redirect('/bill');
     }
